Create sriov network attachment definition
To use SR-IOV with KubeVirt, one needs to have a network attachment
definition that links back to resourceName used in SR-IOV DP config
file. This is the last missing bit in end-to-end SR-IOV integration.

Note that this PR doesn't allow to configure IPAM parameters for the
new network. I'd like to leave this piece separate from this patch.