Pin sriovdp to specific version
This pin also bumps version to the latest in upstream master. It should
work fine because we already create sriovdp configuration file and the
latest image is not broken anymore. (It was at some point broken by
alpine change, which was the original reason to pin it in
kubevirt-ansible, from where the forked image migrated to this
operator).