Fill kubemacpool range in the fillDefaults function
This PR moves the range generation from the render to the fillDefaults functions.

We need this PR so we are able to apply again the cluster CR.

Before this PR we get an error on the operator that update is not
supported on the kubemacpool because we generate a new range and then
the CR is different from the applied one.