fix upgrade of bridge-marker
In older versions of the operator, we used daemon sets of type
'extensions/v1beta1', later we changed that to 'apps/v1'. Because of this
change, we are not able to seamlessly upgrade using only Update methods.
Following code handles this exception by deleting the old daemon set and
creating a new one.

This workaround should be dropped once we handle upgrades of each
component in their respective modules.

Signed-off-by: Petr Horacek <phoracek@redhat.com>