status_manager: only require single ready pod for deployment
Currently we wait until all the pods in deployment become ready.
That is not necessary, we can turn Ready once at least one pod
in the deployment is able to serve requests.

This is needed, so we don't block when deployment uses leader
election in combination with readiness probe and only one pod is
marked as ready at the time.

Signed-off-by: Petr Horacek <phoracek@redhat.com>