more logging and k8s
There was an issue with previous version of Kubernetes where we
ran out of available IPs during lifecycle tests (with dozens of
pods being created and removed in a short period of time).

This PR is adding more logging information when one of our asserts
fail and it also bumps the used k8s version to 1.17.

Signed-off-by: Petr Horacek <phoracek@redhat.com>