limit versions tested in lifecycle lane
Currently we always test upgrade from all released versions to the master
branch.

That has two major issues:

The upgrade test of more than 20 versions back is ran on every PR. This takes a
lot of time and slows down the development cycle. With this PR we test upgrades
only from releases pinned to HCO on regular PRs. That should be sufficient since
we test the full upgrade on release PRs.

When releasing, we test upgrade from to-be-released version to master. That is
failing since the to-be-released version is not yet accessible in Quay. With
this PR we make sure to exclude this version from the test.

Signed-off-by: Petr Horacek <phoracek@redhat.com>