tests: try to catch progressing state more reliably
There is a race on our CI where we sometimes don't catch Progressing state. That
can happen when the the image of a deployed component is already in registry and
therefore the start of it is swift. With this change, we run configuration
change with a little delay after we start listening for the progressing state.
This can still fail, but our chances are better. Better approach would be to
start a monitor instead of polling, that is however not yet supported in the
controller-runtime Client.

Signed-off-by: Petr Horacek <phoracek@redhat.com>