don't remove CRDs when their comonents are gone (#357)
* don't apply owner reference on CRDs

We want to prevent deletion of CRs created by the user, even during removal of
the operator. Therefore, we don't set ownership reference on CRDs.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* don't remove CRDs when a components is removed

Even after a component deletion, we want to keep the CRDs it brought. This
should help us migrate components under different operators. If a user wants to
remove CRDs too, it has to be done manually.

Signed-off-by: Petr Horacek <phoracek@redhat.com>