keep the order of webhooks while merging (#389)
In the current code, we first transform both the current and the new
list of webhooks into a map of webhooks by names. The problem with that
is that a hashmap does not keep the original order. Due to that, the
result of the merging is inconsistent, causing tests to fail and it
would also cause us to reapply config, even when there is nothing needed
to be changed.

With this patch, we iterate the original list of webhooks without
translating it to a map, to keep the original order.

Signed-off-by: Petr Horacek <phoracek@redhat.com>