Ensure components objects deletion and creation at e2e tests (#393)
* Add 'Secret' to Component definition

Secrets should be part of Component type
because some components deploys it.

Secrets objects should be monitored on create
and delete like other types objects.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add removal check of Secrets objects

Handle Secrets objects removal as part of
CheckComponentsRemoval process, ensuring
there is no secret object left in the cluser.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add check for secret deployment

In order to ensure that secret objects are created and
checked in CheckComponentsDeployment function.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add 'MutatingWebhookConfiguration' to Component definition

MutatingWebhookConfiguration should be part of Component type
because some components deploys it.

MutatingWebhookConfigurationh should be monitored on create
and delete like other objects.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add check for MutatingWebhookConfiguration deployment

It is nessesray to add a check for MutatingWebhookConfiguration
existance at CheckComponentsDeployment function,
in order to ensure creation of MutatingWebhookConfiguration.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* Add removal check for MutatingWebhookConfiguration objects

Handle MutatingWebhookConfiguration objects removal
as part of CheckComponentsRemoval process, ensuring
there is no mutatingWebhookConfiguration  object left
in the cluser.

Signed-off-by: Or Mergi <ormergi@redhat.com>