Fix rbac in CNAO's manifest (#419)
* remove over privileged rbac

Currently the cnao rbac Policy has a segment that elevates the container privileges to all resources, and all verbs.
This is too much for the cnao's needs, and also in turn changes scc to anyuid in cnv/okd clusters.
Hence, the rbac segment is removed in this commit.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* refine rbacs to needed operations of cnao

After removing the all inclusive rbac segment, we re-add the appropriate rbac that the cnao requires to operate.

Signed-off-by: Ram Lavi <ralavi@redhat.com>