merge some upgrade tests
For each release, we test upgrades from every previous release to the
newly created one. That takes a lot of time and is obviously getting
only worse.

This patch merges two of the three upgrade tests into a single one. They
are both just checking the state of the cluster, so we are not risking
too much by merging them.

This should save us 30 % of the time on the suite.

Signed-off-by: Petr Horacek <phoracek@redhat.com>