Introduce rejectOwner annotation
In the current code, we mark all objects created by CNAO (those kept
under data/) as owned by us. As a result, they are garbage collected
upon removal of NetworkAddonsConfig.

This cause complications with objects that we do not want to remove or
that we want to be owned by other objects. The recent example of this
were KubeMacPool secrets that need to be precreated in CNAO in order to
avoid chicken-egg issues, but then handled separately.

In order to provide a generic mechanism to avoid ownership, a new
annotation `networkaddonsoperator.network.kubevirt.io/rejectOwner` was
introduced. When an object is annotated with this, we will not set the
reference.

Signed-off-by: Petr Horacek <phoracek@redhat.com>