Don't set owner reference on KMP secret
Currently, we set owner reference of the secret to NetworkAddonsConfig.
That means that it will be removed upon deletion of that CR.

The issue is that there are KMP controllers recreating this Secret, not
knowing whether the whole setup is being deleted or just the specific
Secret.

In order to make this little clearer, we decided not to own the secret
at all and leave its lifecycle on KMP. We still need to pre-create it
with its dummy values in CNAO to prevent chicken-egg issues.

Signed-off-by: Petr Horacek <phoracek@redhat.com>