Introduce `networkaddonsoperator.network.kubevirt.io/rejectOwner` annotation and drop ownership of KMP secrets (#469)
* Introduce rejectOwner annotation

In the current code, we mark all objects created by CNAO (those kept
under data/) as owned by us. As a result, they are garbage collected
upon removal of NetworkAddonsConfig.

This cause complications with objects that we do not want to remove or
that we want to be owned by other objects. The recent example of this
were KubeMacPool secrets that need to be precreated in CNAO in order to
avoid chicken-egg issues, but then handled separately.

In order to provide a generic mechanism to avoid ownership, a new
annotation `networkaddonsoperator.network.kubevirt.io/rejectOwner` was
introduced. When an object is annotated with this, we will not set the
reference.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* Don't set owner reference on KMP secret

Currently, we set owner reference of the secret to NetworkAddonsConfig.
That means that it will be removed upon deletion of that CR.

The issue is that there are KMP controllers recreating this Secret, not
knowing whether the whole setup is being deleted or just the specific
Secret.

In order to make this little clearer, we decided not to own the secret
at all and leave its lifecycle on KMP. We still need to pre-create it
with its dummy values in CNAO to prevent chicken-egg issues.

Signed-off-by: Petr Horacek <phoracek@redhat.com>

* Don't set owner reference on knmstate secret

Currently, we set owner reference of the secret to NetworkAddonsConfig.
That means that it will be removed upon deletion of that CR.

The issue is that there are knmstate controllers recreating this Secret,
not knowing whether the whole setup is being deleted or just the
specific Secret.

In order to make this little clearer, we decided not to own the secret
at all and leave its lifecycle on knmstate. We still need to pre-create
it with its dummy values in CNAO to prevent chicken-egg issues.

Signed-off-by: Petr Horacek <phoracek@redhat.com>