Previx version tag with "v" (#501)
Go mod expects versions to start with "v". Without that, go mod is able
to fetch the tag, but it interprets it as:
v0.3.1-0.20200811082954-d0b735298e44

With this change, we prefix release tags with "v" while keeping the
version in our CSVs with a naked version to satisfy OLM's expectations.

Signed-off-by: Petr Horacek <phoracek@redhat.com>