add stdout to git-clone process in bumper script
Currently when running the bumper script,
there is no output to stdout when a component is cloned
to the temporary folder. As this may take a short while,
it could be seen as if the script is stuck.

Let's dump the clone command output to stdout,
thus making it more user-friendly.

Signed-off-by: Ram Lavi <ralavi@redhat.com>