update CustomResourceDefinition manifests
In CNAO repo the CRD is created by the manifest-templator,
so generation of the CRD with the operator-SDK API is not straight forward.
Instead we manually update components.go to introduce both v1, v1alpha1 version to the CRD
Also, we remove the use of Spec.Version which is deprecated, and update the CRD to use
the appropriate CNAO version types.

Signed-off-by: Ram Lavi <ralavi@redhat.com>