adding a new shared apis folder that holds shared API
Following operator-SDK migration manual, and since v1alpha
is the same in it's content to the new v1, we create a shared folder
to hold all the shared API-Vesrion types, to be used by v1alpha and v1 in future commits.
Also, as the shared folder is just a helper version folder, it does not register it's schema.

Signed-off-by: Ram Lavi <ralavi@redhat.com>