utils, events: add events-emitter module to emit k8s events
In order to bring CNAO to meet operators conventional best practices,
It should use k8s events to describe actions that are being taken.
In this commit a new event emitter is introduced, to easily produce events.
the emitter has to be initiated and then can be used to create events.

Right now the new emitter is aimed to be used only to emit events for configs,
but can be improved in the future to also support associating events to
CSVs and the self pod.

Signed-off-by: alonSadan <asadan@redhat.com>