Migrate CNAO to v1 (#511)
* generate v1 API Version with Operator-SDK

Using operator-SDK API, we generate a new API-Version v1 files
This commit only includes the generated fils with no other changes.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* adding a new shared apis folder that holds shared API

Following operator-SDK migration manual, and since v1alpha
is the same in it's content to the new v1, we create a shared folder
to hold all the shared API-Vesrion types, to be used by v1alpha and v1 in future commits.
Also, as the shared folder is just a helper version folder, it does not register it's schema.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* update empty v1 and v1alpha1's API-Version types to use shared types

We want to reuse shared API-Version type structs to be used both on v1alpha1 and v1.
This way both API-Versions use the same structs via aliasing

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* generate gen-k8s deepcopy files

Using operator-SDK to generate deepcopy files for the v1, v1alpha1, shared API-Versions.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* update CustomResourceDefinition manifests

In CNAO repo the CRD is created by the manifest-templator,
so generation of the CRD with the operator-SDK API is not straight forward.
Instead we manually update components.go to introduce both v1, v1alpha1 version to the CRD
Also, we remove the use of Spec.Version which is deprecated, and update the CRD to use
the appropriate CNAO version types.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* update pkg code to use appropriate cnao library

We need to change all existing CNAO code to use the shared
API-Version types.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* update test code to use appropriate cnao library

We need to update all tests code to use the shared CNAO
API-Version type.

Signed-off-by: Ram Lavi <ralavi@redhat.com>