controller, error-wrap: remove error formatting from Wrapf
When using `Wrapf` it is not necessary to format
the error to the message. It does that automatically.
With the error formatting it duplicates the error message, and
this is of course not needed.

An example (before the change):

cluster  Operator failed: failed to validate NetworkConfig.Spec: invalid configuration:
failed to parse rangeStart because the mac address is invalid: invalid configuration:
failed to parse rangeStart because the mac address is invalid

Signed-off-by: alonSadan <asadan@redhat.com>