Add events to status manager and controller (#515)
* utils, events: add events-emitter module to emit k8s events

In order to bring CNAO to meet operators conventional best practices,
It should use k8s events to describe actions that are being taken.
In this commit a new event emitter is introduced, to easily produce events.
the emitter has to be initiated and then can be used to create events.

Right now the new emitter is aimed to be used only to emit events for configs,
but can be improved in the future to also support associating events to
CSVs and the self pod.

Signed-off-by: alonSadan <asadan@redhat.com>

* status-manager, events: Add k8s events to log deployment status

Add events to log the deployment of all (not each) component.
Three types of events will be logged:  Available, Progressing and Failed.
The events will be associated with the networkAddonsConfig object.

Signed-off-by: alonSadan <asadan@redhat.com>

* Readme, events: add info about events to the readme

Added some info about the newly added events, and how to get them to the readme.

Signed-off-by: alonSadan <asadan@redhat.com>

* controller, error-wrap: remove error formatting from Wrapf

When using `Wrapf` it is not necessary to format
the error to the message. It does that automatically.
With the error formatting it duplicates the error message, and
this is of course not needed.

An example (before the change):

cluster  Operator failed: failed to validate NetworkConfig.Spec: invalid configuration:
failed to parse rangeStart because the mac address is invalid: invalid configuration:
failed to parse rangeStart because the mac address is invalid

Signed-off-by: alonSadan <asadan@redhat.com>

* tests, events: add tests framework for event emitter

- Added a new module for watchhing config Events. This is almost Identical
  to the framwork from kubevirt/kubevort/tests/utils.go
- added checks for all types of config events, to verify they were emitted.
  Can be used in tests, to verify deployment status.

Signed-off-by: alonSadan <asadan@redhat.com>

* tests, events:  added checks for events

added some checks for the existing cases, in order to check
events were emitted properly and that deployment, recovery
and validation works as expected

Signed-off-by: alonSadan <asadan@redhat.com>

* events: run go fmt

keep the go fmt changes in a seperate commit

Signed-off-by: alonSadan <asadan@redhat.com>