Check cnao pod restarts in CI tests (#551)
* add check that cnao pod has not restarted

Currently there is no chck in the tests that
the CNAO pod has gone through a restart.
Let add a check to catch those restarts as soon as they occur.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* add cnao pod stability check to workflow tests

We want to make sure that CNAO pod does not perform
unwanted restart.

To do that we add a consistency check at the end of workflow
test suite to make sure no restart have been made
during the tests.
Also , after each test we want to add a print to signal
where the pod started restarting

Signed-off-by: Ram Lavi <ralavi@redhat.com>