fix flaky in workflow validation test
Due to a flakiness in which the config's Available status is
not changed immediately after a component's removal
we can catch the old Availalbe state and consequently
fail the the validation test.
Due to this, we first wait for the removal changes
to take effect in order to make sure we check
the correct stable statuses.

Signed-off-by: Ram Lavi <ralavi@redhat.com>