bump CustomResourceDefinition to v1 (#537)
* bump CustomResourceDefinition to v1

in order to run "kubectl explain crd"
we need to bump CustomResourceDefinition to v1.
Part of this change is to move the deprecated spec.validation and Subresources.status fields
to be per version in spec.versions[].schema

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* Add minimal crd validation schema

In order for the crd to be validated,
we need to add the minimal schema, including
the cnao components.
This change also includes changing lastHeartbeatTime, lastTransitionTime values to be allow
null values, since in previous versions these values were not declared.
Signed-off-by: Ram Lavi <ralavi@redhat.com>