refactor fix of flaky validation test (#574)
Due to a problem catching the time where a component
was removed from the spec in validation test
the wait for the available status was sometimes missed,
causing a flakiness in the test.
Currently this is solved by a sleep.
Let's improve this by waiting for the specific component removal

Signed-off-by: Ram Lavi <ralavi@redhat.com>