kmp functests, skip KMP opt-mode tests not maching CNAO config
KMP functests include tests for both opt-in and opt-out KMP webhook modes
and since CNAO deploys KMP with opt-in mode,
it is necessary to skip tests of the opposite mode (opt-out tests).

This commit adds a function that checks what is the current configuration of
KMP in the cluster and skip the relevant tests accordingly.

Signed-off-by: Ram Lavi <ralavi@redhat.com>