Add isPrAlreadyOpen method (#507)
* move isBumpNeeded to component_commands.go

We want to add a new function isPrAlreadyOpen to the isBumpNeeded function.
This new function will require githubApi access, so there is no longer
a reason to keep isBumpNeeded outside component_commands.go.
This commit simply moves the functions and unit tests to the appropriate place in component_commands.go.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* adding new function isPrAlreadyOpen to isBumpNeeded

we want to avoid opening multiple bump PRs to the same component's new release.
Let's add a new functio to check that there is no already open PR
with the same bump title.

Signed-off-by: Ram Lavi <ralavi@redhat.com>