Change upgrade test (#569)
* change upgrade test to also update crd manifest

Currently upgrade test only updates operator.yaml manifest,
while keeping the currently deployed crd.
This is not how OLM performs upgrade, and since we want to
test the actual upgrade scenario then we should also upgrade the crd

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* remove UninstallRelease from upgrade test

Currently manifests are updated by
uninstalling the old and intalling the new.
Now that we are also uninstalling the crd,
we need to change this behavior to only update the manifests

Also, since we also remove the CRD during UninstallRelease,
we cant use it in upgrade_test.go test It:"stays in Available
condition while the operator is being removed and redeployed"
So we create an alternate UninstallOperator/InstallOperator
functions to keep test functionality

Signed-off-by: Ram Lavi <ralavi@redhat.com>