bumper, Add bumper make target (#592)
Currently running bumper script requires 2 commands.
The bumper exec output is added to the main directory
and may be mistakenly added to the project.
Let's change the command to run with 'go run',
and move it to a Makefile target for better usability
README is updated accordingly.
Also we add a link to how to produce a github token in README

Also added a link to token creation in case
of running script manually

Signed-off-by: Ram Lavi <ralavi@redhat.com>