Fix installation of bridge and tuning CNIs
We have two serious bugs in the bridge CNI.

One, we never installed our own bridge and tuning CNI, since the source
of our copy was not matching directories used in the CNI plugins
container image. The only reason why this worked so far is that we were
able to consume these plugins from the Multus deployment.

Second, the symlink making sure that we always have an alias to these
CNIs prefixed with cnv- was not working on OpenShift. The reason is that
on OpenShift, the directory containing CNI binaries is different, but
our `ln` did not respect that.

Signed-off-by: Petr Horacek <phoracek@redhat.com>