crdCleanUp, Cleaning release after upgrade test (#587)
Currently, after cnao is uninstalled, the crd's
of it's components are not deleted.
This causes a "downgrade" situation when
upgrade test move from 1 test to the other.
For example, in the first test, release is upgraded
from old version to current version, but in the next test
the old version is not been able to be set because the new crd's
are stil there.

In order to solve this, and since CNAO is not supprting downgrades,
we are attempting to remove the crds manually after each upgrade test.

Signed-off-by: Ram Lavi <ralavi@redhat.com>