bumper, Fix flaky in bumper unit test (#611)
In unit test checking getUpdatedReleaseInfo
entries sometimes failed. This is because
getUpdatedReleaseInfo assumes that the tags list
received by ListMatchingRefs is sorted by time inserted.
However fake ListMatchingRefs method doesn't explicitly
sort it, so sometimes we get different order that causes
the test to fail.
Let's sort the tag list to fix this.
A comment to keep the tags to be inserted in
alphabetic-chronological order is also added.

Signed-off-by: Ram Lavi <ralavi@redhat.com>