bumper, Move config parsing actions to cnaoRepo instance
Let's move all actions regarding components.yaml (aka config file)
to run on the cnaoRepo instance.
To do so, we change the config-path input param to be
relative to the cnaoRepo, so that it'll be easy to concat it
to the cnaoRepo instance.
We also add another reading of the config file to
refresh it's content after each bump.

Signed-off-by: Ram Lavi <ralavi@redhat.com>