bumper, Add resetRepoInAllowedList to remove bump outputs
This method will be performed on the cnaoRepo instance
between each bumping of a component, to remove any bumping residue
Since bump output files are located in specific locations,
we add an allowList, to make sure only the bumping outputs
get removed.
Also adding a final resetRepoInAllowedList after all components
got bumped to finish the bumper script with no residual files.
in order to use the resetRepo function with defer, we move the
bump operation to a function.

Signed-off-by: Ram Lavi <ralavi@redhat.com>