bumper, Add bumping repo instance (#616)
* bumper, Add a cnao repo instance and cnaoRepoCommands.go

Currently, functions that run on the cnao repo
are run with the component api, instead of on it's own instance.
To make things neater, Let's create a cnaoRepo instance,
and use it to run cnao repo related actions.
Additionaly, all cnao repo related function move to cnaoRepoCommands.go

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Move cnap repo related tests to cnaoRepoCommands_test.go

We purely* move the unit-tests related to cnaoRepo to the dedicated
test file cnaoRepoCommands_test.go

* only exception is a small fix in isComponentBumpNeeded where we also
update the url so that the test won't fail.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Move config parsing actions to cnaoRepo instance

Let's move all actions regarding components.yaml (aka config file)
to run on the cnaoRepo instance.
To do so, we change the config-path input param to be
relative to the cnaoRepo, so that it'll be easy to concat it
to the cnaoRepo instance.
We also add another reading of the config file to
refresh it's content after each bump.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Move running  bump-<component> scripts to cnaoRepo instance

Let's move bump component script actions to cnaoRepo instance

Signed-off-by: Ram Lavi <ralavi@redhat.com>