cluster scripts, Fix ambiguous between cluster and _kubevirtci folders
Currently, both _kubecirtci/ and cluster/ scripts
use the KUBEVIRTCI_PATH var. The issue is that they
point to different paths.
This was not an issue until now because they are only
used internally.
However if one tries to set this variable and calls a script
that use both scripts (for example cluster/kubectl.sh) then the
script breaks due to this mismatch.
kubevirtci.sh script namings further contibute to this ambiguity.
- Changed cluster/ scripts to use different variable name: CLUSTER_PATH
- Changed kubevirtci.sh script name to cluster.sh
- Fixed script function names accordingly: cluster::<function-name>

Signed-off-by: Ram Lavi <ralavi@redhat.com>