delete objects, Fix tracking of already removed objects (#665)
* delete objects, Fix tracking of already removed objects

Currently, the reconciler updates the list of objects to deploy
only after any objects are removed. This causes a temporary period
where the object is declared as gone.
Changed the order of trackDeployedObjects and deleteObjects to
keep the tracked objects updated correctly.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* expectAvaialableEvent, Change to Watch policy to since now

In order to check Available event after object removal,
we move to use CheckAvailableEvent.
We also change the event watch policy to SinceNow,
since it seems that the change is sufficiently caught.

Signed-off-by: Ram Lavi <ralavi@redhat.com>