[release-0.42] backport bumper commits to release-0.42 branch (#687)
* bumper, Add bumper make target (#592)

Currently running bumper script requires 2 commands.
The bumper exec output is added to the main directory
and may be mistakenly added to the project.
Let's change the command to run with 'go run',
and move it to a Makefile target for better usability
README is updated accordingly.
Also we add a link to how to produce a github token in README

Also added a link to token creation in case
of running script manually

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Add go install make target dependency to auto-bumper in Makefile (#625)

Currently auto-bumper will fail if go is not installed
Added the go install target dependency to fix that

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Fix macvtap bump script name (#646)

Currently the macvtap bump script name "bump-macvtap.sh"
does not match the name in components.yaml.
In order to easily call it from the bumper script
Changed the name to "bump-macvtap-cni.sh"

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* bumper, Adapt nmstate bump script to comply with components.yaml name

currently nmstate bump script is called bump-knmstate.sh
Let's change it to bump-nmstate.sh to comply with
components.yaml.
This way it's would be easier to call this script from the bumper script.

Signed-off-by: Ram Lavi <ralavi@redhat.com>