bumper, fix returning annotated tag sha on tagged update-policy (#709)
Currently when using tagged update policy mode, the bumper script is returning
the annotated tag sha and not the relevant commit sha of this tag.
This will cause the future bump to fail, as the annotated tag is directly connected
via the githubApi.
Moreover, the mocked listMatchingRefs routine did not catch it since it used git log
in order to assemple the tag list.
Fixed to return the commit sha.
Fixed the listMatchingRefs mock to use 'show-ref --tag', and adjusted the test to catch this issue.

Signed-off-by: Ram Lavi <ralavi@redhat.com>