bumper, Change bumper fail policy (#708)
Currently, the bumper script will fail upon the first bump failure.
We want to be able to run the bumper on all components even if one
component bump is broken.
The new behavior is that all found bumps will occur, and in the end if at least one
component failed to bump - the job will exit with error.

Signed-off-by: Ram Lavi <ralavi@redhat.com>