Migrate NMState to use kubernetes-nmstate-operator if available
This patch make CNAO use the kubernetes-nmstate-operator if it is
running and available. The only prereq for this is to have the
operator pod running. If that's detected via the
app=kubernetes-nmstate-operator label on a Deployment, then CNAO
will only render an nmstates.nmstate.io CR to kick off the handler
and webhook from the operator.

If the operator is added, it also has upgrade logic to move to using
the operator instead of installing kubernetes-nmstate directly. It
will remove the old instance.

Signed-off-by: Brad P. Crochet <brad@redhat.com>