Use default at bad old self signed conf (#813)
If previous selfSignConfiguration is missing a field like in the case of
certOverlapInterval (it was not introduced at the beginning) then the
default values are not going to be used and the defaults from the
manifests apply so we can end up with a bad overall configuration. This
change use default values if previous ones were missing one of the
fields.

Signed-off-by: Quique Llorente <ellorent@redhat.com>