checkForDeployment, don't check on CNAO operator itself (#942)
The CNAO Operator does not need to have the version label (it
already has it on annotation).
The reason CI passed up until now is that the deployment had no labels
and the test didn't fail this scenario.
Fixing to check only needed cases (when checking CNAO components' deployments)
and returning error if labels are non-existing.

Signed-off-by: Ram Lavi <ralavi@redhat.com>