data/linux-bridge, Fix installation of bridge and tuning CNIs
There is a bug where the linux-bridge pods never installed the bridge
and tuning CNI, since the source address of the copy was not matching
directories used in the CNI plugins container image.
The only reason why this worked so far is that these plugins are
deployed by other CNI intallations used by kubevirtci cluster-up.

Signed-off-by: Ram Lavi <ralavi@redhat.com>