components: Consume CNI Bridge with macspoofchk support (#945)
* components: Consume CNI Bridge with macspoofchk support

As an intermediate and temporary step, consume the changes posted to the
CNI bridge [1].

Once the PR is merged into the original repo, this change will be
reverted.

In case the PR will not be merged into the original repo, we will
proceed with a more formal resolution.

[1] https://github.com/containernetworking/plugins/pull/639

Signed-off-by: Edward Haas <edwardh@redhat.com>

* components: Build linux-bridge CNI from new sources

This patch builds a new image from the newly updated components.yaml.

Signed-off-by: Petr Horáček <phoracek@redhat.com>

Co-authored-by: Petr Horáček <phoracek@redhat.com>