nmstate-webhook:set num of replicas based on infra
When running on a singleReplica openshift cluster, CNAO should run
a single replica of each of It's operands. Currently the number of nmstate-webhook replicas
is hardcoded to 2.

This commit set the number of nmstate-webhook replicas dynamically
depending on the cluster defined infrastructure topology:
SingleReplica(1) or HighAvailable(2).

Also this commit set the number of minimum replicas of the webhook pod to 0 on SNO,
otherwise upgardes won't work because when the pod will be upgrade the number of
active pods will be 0.

Signed-off-by: alonsadan <asadan@redhat.com>