Determine webhook replicas based on infra (#973)
* Bump openshift/api for cluster HA mode API

Bump to github.com/openshift/api@670ac3fc997c2f1d19b8c29ef04f70d6e3d4a59e
to get the infrastructure topology API.

Signed-off-by: alonsadan <asadan@redhat.com>

* nmstate-webhook:set num of replicas based on infra

When running on a singleReplica openshift cluster, CNAO should run
a single replica of each of It's operands. Currently the number of nmstate-webhook replicas
is hardcoded to 2.

This commit set the number of nmstate-webhook replicas dynamically
depending on the cluster defined infrastructure topology:
SingleReplica(1) or HighAvailable(2).

Also this commit set the number of minimum replicas of the webhook pod to 0 on SNO,
otherwise upgardes won't work because when the pod will be upgrade the number of
active pods will be 0.

Signed-off-by: alonsadan <asadan@redhat.com>