ReconcileNetworkAddonsConfig, Remove redundant call to SetFromPods()
SetFromPods sets the status depending on the current components objects
deployed. Although originally used on pod-controller, we call it here as
well to have quicker status updates.
The issue is that every ReconcileNetworkAddonsConfig controller
reconcile (that happens every 1 minute), we call SetFromPods twice. This
is redundant.
Run SetFromPods() once after trackDeployedObjects so the components are
updated.

Signed-off-by: Ram Lavi <ralavi@redhat.com>