(BZ#2006810) status_manger, Reduce number of status update API calls (#1032) (#1063)
* ReconcileNetworkAddonsConfig, Remove redundant call to SetFromPods()

SetFromPods sets the status depending on the current components objects
deployed. Although originally used on pod-controller, we call it here as
well to have quicker status updates.
The issue is that every ReconcileNetworkAddonsConfig controller
reconcile (that happens every 1 minute), we call SetFromPods twice. This
is redundant.
Run SetFromPods() once after trackDeployedObjects so the components are
updated.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pod_controller, Remove RequeueAfter at Reconciler

This controller watch for component changes asynchronously, hence there
is no real need to resync these requests.
Moreover, SetFromPods is already called synchronously from
ReconcileNetworkAddonsConfig controller.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* status_manager, Aggregate status update when possible

On non-failing routine status updates, it is better to aggregate the
status conditions in order to reduce the amount of status update API
calls.
This reduces the amount of status updates per
ReconcileNetworkAddonsConfig controller run to 2 (one for operator-level
status and one for Pod-Deployment-level)

Signed-off-by: Ram Lavi <ralavi@redhat.com>