cluster-info: add info on podSecurity feature
podSecurity feature allows to enforce pod security standarts[1]
as an admission controller. This feature is enabled
in the cluster level (and then applied in the namesapce level), and it not
necessarily enabled by default (by the time of this commit it is only
exists on k8s-1.22 and it is not enabled by default).
It will be useful to know if it can be applied on the components or not.

This commit adds a flag to the cluster-info to help with that.

https://kubernetes.io/docs/concepts/security/pod-security-standards/

Signed-off-by: alonsadan <asadan@redhat.com>